Interface ChronusQ GIAO overlap integrals.
