+++ gqcp/include/Basis/Integrals/Interfaces/ChronusQ/CMakeLists.txt
@@ -1,4 +1,7 @@
 target_sources(gqcp
     PRIVATE
+        aliases.hpp
         engines.hpp
+        Environment.hpp
+        math.hpp
 )
