+++ gqcp/include/Basis/Integrals/Interfaces/ChronusQ/Environment.hpp
@@ -27,30 +27,28 @@
 
 #pragma once
 
-#include "Basis/Integrals/Interfaces/ChronusQ/aliases.hpp"
-#include "Basis/Integrals/Interfaces/LibintInterfacer.hpp"
 
+#include <array>
 #include <vector>
 
 
 namespace ChronusQ {
 
 
-struct ComplexGIAOIntEngine {
+/**
+ *  A wrapper around ChronusQ's global variables.
+ */
+struct Environment {
 
-    static std::vector<std::vector<dcomplex>> computeGIAOOverlapS(libint2::ShellPair&, libint2::Shell&, libint2::Shell&, const std::array<double, 3>&);
+    static std::vector<std::vector<std::array<int, 3>>> cart_ang_list;
+    static std::vector<std::vector<std::array<int, 3>>> pop_cart_ang_list();
 
-    // calculate the uncontracted overlap of (s||s) type for a shellpair
-    static std::vector<dcomplex> computecompOverlapss(libint2::ShellPair&, libint2::Shell&, double*, libint2::Shell&, double*);
+    static std::vector<std::vector<double>> car2sph_matrix;
+    static std::vector<std::vector<double>> pop_car2sph_matrix();
 
-    // complex overlap horizontal recursion for contracted case
-    static dcomplex comphRRSab(libint2::ShellPair&, libint2::Shell&, libint2::Shell&, double*, std::vector<dcomplex>&, int, int*, int, int*);
 
-    // complex overlap horizontal recursion iPP specific for uncontracted case
-    static dcomplex comphRRiPPSab(libint2::ShellPair::PrimPairData&, libint2::Shell&, libint2::Shell&, double*, dcomplex, int, int*, int, int*);
-
-    // complex overlap vertical recursion for uncontracted case
-    static dcomplex compvRRSa0(libint2::ShellPair::PrimPairData&, libint2::Shell&, double*, dcomplex, int, int*);
+    static std::array<std::array<double, 25>, 3201> FmTTable;
+    static std::array<std::array<double, 25>, 3201> generateFmTTable();
 };
