+++ gqcp/include/gqcp.hpp
@@ -24,6 +24,12 @@
 #include "Basis/Integrals/BaseTwoElectronIntegralEngine.hpp"
 #include "Basis/Integrals/IntegralCalculator.hpp"
 #include "Basis/Integrals/IntegralEngine.hpp"
+#include "Basis/Integrals/Interfaces/ChronusQ/Environment.hpp"
+#include "Basis/Integrals/Interfaces/ChronusQ/aliases.hpp"
+#include "Basis/Integrals/Interfaces/ChronusQ/engines.hpp"
+#include "Basis/Integrals/Interfaces/ChronusQ/math.hpp"
+#include "Basis/Integrals/Interfaces/ChronusQOneElectronIntegralBuffer.hpp"
+#include "Basis/Integrals/Interfaces/ChronusQOneElectronIntegralEngine.hpp"
 #include "Basis/Integrals/Interfaces/LibcintInterfacer.hpp"
 #include "Basis/Integrals/Interfaces/LibcintOneElectronIntegralBuffer.hpp"
 #include "Basis/Integrals/Interfaces/LibcintOneElectronIntegralEngine.hpp"
