+++ gqcp/src/Basis/Integrals/Interfaces/ChronusQ/CMakeLists.txt
@@ -1,4 +1,6 @@
 target_sources(gqcp
     PRIVATE
         engines.cpp
+        Environment.cpp
+        math.cpp
 )
